rag_variants/requirements.txt
python-dotenv
tiktoken
chromadb
loguru
tavily-python
bs4
huggingface_hub
grandalf
matplotlib
pexpect
langchain
langchain-mistralai
langchain-groq
langchain_community
langchainhub
langgraph
exa_py
crewai
crewai[tools]
